+++ src/allotropy/allotrope/schemas.py
@@ -1,20 +1,21 @@
 import json
 import os
-from pathlib import Path
-import re
 from typing import Any
 
-SCHEMAS_DIR = os.path.join(Path(__file__).parent, "schemas")
-SHARED_SCHEMAS_DIR = os.path.join(SCHEMAS_DIR, "shared", "definitions")
+from allotropy.allotrope.schema_parser.path_util import (
+    get_schema_path_from_manifest,
+    SCHEMA_DIR_PATH,
+    SHARED_SCHEMAS_PATH,
+)
 
 
 def get_shared_definitions() -> dict[str, Any]:
-    with open(os.path.join(SHARED_SCHEMAS_DIR, "definitions.json")) as f:
+    with open(os.path.join(SHARED_SCHEMAS_PATH, "definitions.json")) as f:
         return json.load(f)  # type: ignore[no-any-return]
 
 
 def get_shared_unit_definitions() -> dict[str, Any]:
-    with open(os.path.join(SHARED_SCHEMAS_DIR, "units.json")) as f:
+    with open(os.path.join(SHARED_SCHEMAS_PATH, "units.json")) as f:
         return json.load(f)  # type: ignore[no-any-return]
 
 
@@ -25,7 +26,7 @@ def add_definitions(schema: dict[str, Any]) -> dict[str, Any]:
         ("custom", "custom"),
     ]:
         existing = schema.get(f"${section}", {})
-        with open(os.path.join(SHARED_SCHEMAS_DIR, f"{file}.json")) as f:
+        with open(os.path.join(SHARED_SCHEMAS_PATH, f"{file}.json")) as f:
             additional = json.load(f)
         additional.update(existing)
         schema[f"${section}"] = additional
@@ -33,16 +34,8 @@ def add_definitions(schema: dict[str, Any]) -> dict[str, Any]:
 
 
 def get_schema(schema_relative_path: str) -> dict[str, Any]:
-    with open(os.path.join(SCHEMAS_DIR, schema_relative_path)) as f:
+    with open(os.path.join(SCHEMA_DIR_PATH, schema_relative_path)) as f:
         return add_definitions(json.load(f))
-
-
-def get_schema_path_from_manifest(manifest: str) -> str:
-    match = re.match(r"http://purl.allotrope.org/manifests/(.*)\.manifest", manifest)
-    if not match:
-        msg = f"No matching schema in repo for manifest: {manifest}"
-        raise ValueError(msg)
-    return f"{match.groups()[0]}.json"
 
 
 def get_schema_from_manifest(manifest: str) -> dict[str, Any]: