+++ src/allotropy/parsers/agilent_gen5_image/agilent_gen5_image_parser.py
@@ -35,6 +35,7 @@ from allotropy.parsers.agilent_gen5_image.agilent_gen5_image_structure import (
 )
 from allotropy.parsers.agilent_gen5_image.constants import (
     DEFAULT_SOFTWARE_NAME,
+    DETECTION_TYPE,
     DEVICE_TYPE,
     MULTIPLATE_FILE_ERROR,
     NO_PLATE_DATA_ERROR,
@@ -173,7 +174,7 @@ class AgilentGen5ImageParser(VendorParser):
                             device_control_document=[
                                 OpticalImagingDeviceControlDocumentItem(
                                     device_type=DEVICE_TYPE,
-                                    detection_type=read_section.image_mode.value,
+                                    detection_type=DETECTION_TYPE,
                                     # This setting won't get reported at the moment since Gen5 only reports it
                                     # in micrometers and we don't do conversions on the adapters at the moment
                                     # detector_distance_setting__plate_reader_=get_instance_or_none(