+++ CHANGELOG.md
@@ -9,14 +9,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Security
+
+## [0.1.32]
+
+### Added
+
+- Add schema_parser/path_util.py to remove dependency: converter.py -> generate_schemas.py, which pulled script dependencies into allotropy
+
+### Fixed
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Security
+
+## [0.1.31] - 2024-05-24
+
+### Added
+
 - Script to create graph visualization of calculated data documents from asm json files
 - Details of parser requirements to docs
 - Add Agilent Gen5 Image Adapter
+- Add CTL Immunospot adapter
 - Add Methodical Mind adapter
 
 ### Fixed
 
 - Fixed missing required field in cell-counting 2023/11 schema
+- Fixed missing required field in 2023/09 lum/fluor/abs plate reader schemas
 
 ### Changed
 - Upgraded allotropy python requirement to python 10